Added Implementation for Staff Tab
Created a Database method to retrieve the Staff data from the SAKILA database. Added a method in the Database class to to process this data so that it is ready to be used in the app. Added a method, with additional helper methods, in the App class to display staff data in the App.