Fixed Staff pane and added Inventory Pane
Changed the store_id column to Store in the Staff pane.
Changed the query to show the store name instead of the store ID in the Staff pane.
Added a method to the Database class to retrieve a summary of the inventory of each store.
Added a method to the App class to add the retrieved summary data of each store to the Inventory pane.