fix: Changed staff filter to provide more options
Changed the staff filter so that the table can be filtered by a specific column.
Added the staff ID column to the staff pane and the category column to the film pane.
Added custom upNumber tag to the pom.xml file. Commented it out for now because it is creating errors.